Move AccuWeatherSensorDescription and SENSOR consts to sensor platform
